Get All Customers Comleted - still shows non-customers
