+++ Controllers/UserProfileController.cs
@@ -61,76 +61,6 @@ public class UserProfilesController : ControllerBase
         });
     }
 
-    // Get Customer Profile by Id
-    [HttpGet("customer/{id}")]
-    // [Authorize]
-    public IActionResult GetCustomerById(int id)
-    {
-        var customer = _dbContext
-            .UserProfiles
-                .Include(u => u.IdentityUser)
-            .SingleOrDefault(c => c.Id == id);
-
-        var orders = _dbContext
-            .Orders
-                .Include(o => o.Employee)
-                    .ThenInclude(e => e.IdentityUser)
-                .Include(o => o.OrderItems)
-                    .ThenInclude(oi => oi.AppleVariety)
-            .Where(o => o.CustomerUserProfileId == id);
-
-        if (customer == null)
-        {
-            return NotFound();
-        }
-
-        return Ok(new CustomerDTO
-        {
-            Id = customer.Id,
-            FirstName = customer.FirstName,
-            LastName = customer.LastName,
-            Address = customer.Address,
-            Email = customer.IdentityUser.Email,
-            Orders = orders.Select(o => new OrderDTO
-            {
-                Id = o.Id,
-                CustomerUserProfileId = o.CustomerUserProfileId,
-                Customer = null,
-                EmployeeUserProfileId = o.EmployeeUserProfileId,
-                Employee = null,
-                // Employee = new OrderPickerDTO
-                // {
-                //     Id = o.Employee.Id,
-                //     FirstName = o.Employee.FirstName,
-                //     LastName = o.Employee.LastName,
-                //     Address = o.Employee.Address,
-                //     Email = o.Employee.IdentityUser.Email,
-                //     CompletedOrders = null
-                // },
-                DateOrdered = o.DateOrdered,
-                DateCompleted = o.DateCompleted,
-                Canceled = o.Canceled,
-                OrderItems = o.OrderItems.Select(oi => new OrderItemDTO
-                {
-                    Id = oi.Id,
-                    OrderId = oi.OrderId,
-                    AppleVarietyId = oi.AppleVarietyId,
-                    AppleVariety = new AppleVarietyDTO
-                    {
-                        Id = oi.AppleVariety.Id,
-                        Type = oi.AppleVariety.Type,
-                        ImageUrl = oi.AppleVariety.ImageUrl,
-                        CostPerPound = oi.AppleVariety.CostPerPound,
-                        IsActive = oi.AppleVariety.IsActive,
-                        Trees = null,
-                        OrderItems = null
-                    },
-                    Pounds = oi.Pounds
-                }).ToList()
-            }).ToList()
-        });
-    }
-
     // Get UserProfiles with Roles
     [HttpGet("withroles")]
     [Authorize(Roles = "Admin")]
